# iscc_hub/static/openapi.yaml
openapi: 3.1.0
info:
  title: ISCC-HUB API
  description: |
    Sign, timestamp, and discover content with ISCC-IDs

    **Important**: This API uses content negotiation. All requests must include the
    `Accept: application/json` header to receive JSON responses. Requests without this
    header will be routed to HTML views instead of the JSON API.
  version: 0.1.0
  x-logo:
    url: "/img/iscc-logo-white.png"
    backgroundColor: "#0054b2"
    altText: "International Standard Content Code (ISCC) - Logo"
servers:
  - url: http://localhost:8000
    description: Development server (requires Accept application/json header)
paths:
  /lookup:
    get:
      summary: Lookup Declarations
      description: |
        Exact-match lookup of ISCC declarations by datahash or ISCC-CODE.

        **Lookup Parameters:** Provide exactly ONE of the following parameters:
        - `datahash` - Find all declarations for specific content (by blake3 hash)
        - `iscc_code` - Find all declarations with a specific ISCC-CODE

        **Important:** Providing both parameters or neither will return a 400 error.

        **Use Cases:**
        - After calculating a file's blake3 hash client-side, construct the datahash
          (`1e20` + blake3_hex) and query for existing declarations
        - Lookup by ISCC-CODE to find all timestamped instances of that code

        **Note:** Multiple declarations (ISCC-IDs) may exist for the same datahash or
        ISCC-CODE if content was declared multiple times.

        For similarity search (approximate matching), see `GET /search` and
        `POST /search`.
      operationId: lookup
      parameters:
        - name: datahash
          in: query
          required: false
          schema:
            type: string
            pattern: "^1e20[0-9a-f]{64}$"
            minLength: 68
            maxLength: 68
          description: |
            Blake3 multihash of the content to search for.

            **Format:** `1e20` (blake3 prefix) + 64 hex characters
            **Mutually exclusive with:** `iscc_code`
          example: "1e205ca7815adcb484e9a136c11efe69c1d530176d549b5d18d038eb5280b4b3470c"
        - name: iscc_code
          in: query
          required: false
          schema:
            type: string
            pattern: "^ISCC:[A-Z2-7]{29,68}$"
            minLength: 34
            maxLength: 73
          description: |
            ISCC-CODE to search for.

            **Format:** `ISCC:` followed by 29-68 base32 characters
            **Mutually exclusive with:** `datahash`
          example: "ISCC:KACWN77F73NA44D6EUG3S3QNJIL2BPPQFMW6ZX6CZNOKPAK23S2IJ2I"
      responses:
        '200':
          description: Successful lookup - returns matching declarations (may be empty array)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IsccDeclaration'
              examples:
                lookup_by_datahash:
                  summary: Lookup by datahash - multiple declarations for same content
                  value:
                    - iscc_id: "ISCC:MAIGHFECJMOPMIAB"
                      iscc_code: "ISCC:KED572P4AOF5K6QXQA4T6OJD5UGX7UBPFW2TVQNTHBCKFRFCANCZARQ4K6NSFZQSH4GQ"
                      datahash: "1e205ca7815adcb484e9a136c11efe69c1d530176d549b5d18d038eb5280b4b3470c"
                      timestamp: "2025-01-20T14:32:18.123456Z"
                      controller: "did:web:example.com"
                      pubkey: "z6MkmeDbeC5BecFmVnTHA5PWEBaVUrGLdB3weGE2KYnXfHso"
                      gateway: "https://example.com/metadata/maighfecjmopmiab"
                      metahash: "1e202335f74fc18e2f4f99f0ea6291de5803e579a2219e1b4a18004fc9890b94e598"
                    - iscc_id: "ISCC:MABXW23FQLCYPKIZ"
                      iscc_code: "ISCC:KACT3NRHBQXWJKXV6OJ7XR7B62NJXMYQDNHCNZQZLW4L2UXBCJIZERA"
                      datahash: "1e205ca7815adcb484e9a136c11efe69c1d530176d549b5d18d038eb5280b4b3470c"
                      timestamp: "2025-01-19T09:15:42.789012Z"
                      pubkey: "z6MknNWEmX1zYYZbCCjWGYja9gZA64AKrKNLtsdP2g5EkFrB"
                lookup_by_iscc_code:
                  summary: Lookup by ISCC-CODE - single declaration
                  value:
                    - iscc_id: "ISCC:MADKZ5PLNRXGTVQB"
                      iscc_code: "ISCC:KACTXL7PSZSWDVMQ2DPDF6FXHTKLM3Y5WNRQZQAJDKQRTXHIZXMIRJI"
                      datahash: "1e20abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
                      timestamp: "2025-01-18T21:47:33.456789Z"
                      controller: "did:web:archive.example.org"
                      pubkey: "z6MkfrQkTJTQzDNVbj7gPYGZvL89vVBEFfK4zePNKhRMPYBk"
                      gateway: "https://archive.example.org/iscc/madkz5plnrxgtvqb"
                no_results:
                  summary: No declarations found
                  value: []
        '400':
          description: Bad Request - Invalid parameter or mutual exclusivity violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                no_parameters:
                  summary: No search parameter provided
                  value:
                    detail: "Exactly one search parameter required: datahash or iscc_code"
                both_parameters:
                  summary: Multiple search parameters provided
                  value:
                    detail: "Only one search parameter allowed: datahash or iscc_code (not both)"
                invalid_datahash:
                  summary: Invalid datahash format
                  value:
                    detail: "Invalid datahash format. Expected: 1e20 followed by 64 hex characters"
                invalid_iscc_code:
                  summary: Invalid ISCC-CODE format
                  value:
                    detail: "Invalid iscc_code format. Expected: ISCC: followed by 29-68 base32 characters"
  /search:
    get:
      summary: Similarity Search by ISCC-CODE
      description: |
        Find indexed assets similar to the given ISCC-CODE.

        Similarity search is **optional** Hub functionality, proxied to an external
        iscc-search backend holding a hub-global index. Hubs that do not offer it
        respond with `404` to every `/search` request.

        For exact-match lookup of declarations on this Hub, see `GET /lookup`.
      operationId: search
      parameters:
        - name: iscc_code
          in: query
          required: true
          schema:
            type: string
            pattern: "^ISCC:[A-Z2-7]{16,}$"
          description: Composite ISCC-CODE to find similar assets for
          example: "ISCC:KADUHBUDQUT3LPWRJH6BUAG7HMBIXX6JRQRX3JH7EBIOSMXEVL5URBBUPOIOTU4HLSSQ"
        - $ref: '#/components/parameters/SearchLimit'
      responses:
        '200':
          description: Similarity matches, ordered by relevance (may be empty)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsccSearchResult'
        '400':
          $ref: '#/components/responses/SearchBadRequest'
        '404':
          $ref: '#/components/responses/SearchNotFound'
        '422':
          $ref: '#/components/responses/SearchUnprocessable'
        '503':
          $ref: '#/components/responses/SearchUnavailable'
    post:
      summary: Similarity Search by Query
      description: |
        Find indexed assets similar to the given query (`IsccQuery`).

        At least one of `iscc_id`, `iscc_code`, or `units` must be provided.
        `iscc_id` takes precedence (more-like-this against an indexed asset) and
        yields `404` when the reference is not indexed. Top-level body keys outside
        the documented set — including `simprints` — are rejected with `400`.

        Similarity search is **optional** Hub functionality, proxied to an external
        iscc-search backend holding a hub-global index. Hubs that do not offer it
        respond with `404` to every `/search` request.
      operationId: search_by_query
      parameters:
        - $ref: '#/components/parameters/SearchLimit'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IsccQuery'
            examples:
              by_iscc_code:
                summary: Similarity by composite ISCC-CODE
                value:
                  iscc_code: "ISCC:KADUHBUDQUT3LPWRJH6BUAG7HMBIXX6JRQRX3JH7EBIOSMXEVL5URBBUPOIOTU4HLSSQ"
              by_units:
                summary: Similarity by explicit ISCC-UNITs
                value:
                  units:
                    - "ISCC:AAAUHBUDQUT3LPWR"
                    - "ISCC:CAAUT7A2ADPTWAUL"
              more_like_this:
                summary: More-like-this by indexed ISCC-ID
                value:
                  iscc_id: "ISCC:MAIGIIFJRDGEQQAA"
      responses:
        '200':
          description: Similarity matches, ordered by relevance (may be empty)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsccSearchResult'
        '400':
          $ref: '#/components/responses/SearchBadRequest'
        '404':
          $ref: '#/components/responses/SearchNotFound'
        '422':
          $ref: '#/components/responses/SearchUnprocessable'
        '503':
          $ref: '#/components/responses/SearchUnavailable'
  /declaration:
    post:
      summary: Declare ISCC-CODE
      description: |
        Submit an `IsccNote` to declare an ISCC-CODE. The server validates the submission
        and sequences it in the declaration log, then returns a minimal `DeclarationAck`
        with the assigned `iscc_id` and its log `seq`. The full signed `IsccReceipt`
        (W3C Verifiable Credential) is retrieved separately from
        `GET /declaration/{iscc_id}/receipt`, keeping the submission hot path free of
        signing work.

        **Duplicate Protection:** By default, the server rejects duplicate declarations
        (same datahash) with a 409 Conflict error. To force a duplicate declaration,
        include the `X-Force-Declaration: true` header.
      operationId: declare
      parameters:
        - name: X-Force-Declaration
          in: header
          required: false
          schema:
            type: string
            enum: ["true", "1"]
          description: |
            Set to "true" or "1" to bypass duplicate declaration checks and force
            creation of a new ISCC-ID for the same content (datahash).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IsccNote'
            examples:
              full_example:
                summary: Complete IsccNote with all optional fields
                value:
                  $schema: "http://purl.org/iscc/schema/iscc-note-0.8.0.json"
                  iscc_code: "ISCC:KACWN77F73NA44D6EUG3S3QNJIL2BPPQFMW6ZX6CZNOKPAK23S2IJ2I"
                  datahash: "1e205ca7815adcb484e9a136c11efe69c1d530176d549b5d18d038eb5280b4b3470c"
                  metahash: "1e202335f74fc18e2f4f99f0ea6291de5803e579a2219e1b4a18004fc9890b94e598"
                  nonce: "0013a3c214c05796673503e6e549446d"
                  timestamp: "2025-01-15T12:00:00.000Z"
                  gateway: "https://example.com/iscc_id/{iscc_id}/metadata"
                  units:
                  - "ISCC:AADWN77F73NA44D6X3N4VEUAPOW5HJKGK5JKLNGLNFPOESXWYDVDVUQ"
                  - "ISCC:EADSKDNZNYGUUF5AMFEJLZ5P66CP5YKCOA3X7F36RWE4CIRCBTUWXYY"
                  - "ISCC:GAD334BLFXWN7QWLCSBGJMLRZW73FFNV7ORVUKN23UWPKGQCWTIHQKY"
                  signature:
                    version: "ISCC-SIG v1.0"
                    controller: "did:web:example.com"
                    pubkey: "z6MkmeDbeC5BecFmVnTHA5PWEBaVUrGLdB3weGE2KYnXfHso"
                    proof: "z5j9nrpPw3oYSAN4XbCvk2sUtkwrueTD6V2Y35gS1KFTode2ED2YQWokPmoXw6QBYtYEFxtAQfzBhdNyr8PMwP79G"
              minimal_example:
                summary: Minimal IsccNote with only required fields
                value:
                  $schema: "http://purl.org/iscc/schema/iscc-note-0.8.0.json"
                  iscc_code: "ISCC:KACWN77F73NA44D6EUG3S3QNJIL2BPPQFMW6ZX6CZNOKPAK23S2IJ2I"
                  datahash: "1e205ca7815adcb484e9a136c11efe69c1d530176d549b5d18d038eb5280b4b3470c"
                  nonce: "00197932bfe2d31f11d8709be50f84bd"
                  signature:
                    version: "ISCC-SIG v1.0"
                    controller: "did:web:example.com"
                    pubkey: "z6MkmeDbeC5BecFmVnTHA5PWEBaVUrGLdB3weGE2KYnXfHso"
                    proof: "znTcLukuu7CL8FpoaSxr6UWQpxCogZUCZ9iqx9dnP3inkzGszbUN79qrxWvpxEHnuNtLQ5Jd1rBukY7g1JSiYHVZ"
      responses:
        '201':
          description: Declaration successful - ISCC-ID created and sequenced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeclarationAck'
        '400':
          description: Bad Request - General client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid signature or authentication. Returns the first error encountered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Duplicate declaration detected. Use X-Force-Declaration header to override.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity - Validation failed. Returns the first validation error encountered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /declaration/{iscc_id}/receipt:
    get:
      summary: Get IsccReceipt
      description: |
        Retrieve the signed `IsccReceipt` (W3C Verifiable Credential) for a previously
        sequenced declaration, reconstructed from the stored log record. This is the
        verifiable counterpart to the minimal `DeclarationAck` returned by
        `POST /declaration`.
      operationId: get_declaration_receipt
      parameters:
        - name: iscc_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^ISCC:[A-Z2-7]{16}$"
          description: The ISCC-ID of the declaration to build a receipt for
          example: "ISCC:MAIGIEGXNGY664AB"
      responses:
        '200':
          description: The signed IsccReceipt for the declaration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsccReceipt'
        '404':
          description: Not Found - No declaration exists for this ISCC-ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /declaration/{iscc_id}:
    delete:
      summary: Delete ISCC-ID
      description: |
        Delete a previously timestamped ISCC declaration. The deletion request must be signed
        by the same controller that created the original declaration.

        **Side Effects:**
        - A deletion record is appended to the transparency log with a new sequence number
        - The deletion is timestamped with microsecond precision
      operationId: delete_declaration
      parameters:
        - name: iscc_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^ISCC:[A-Z2-7]{16}$"
          description: The ISCC-ID of the declaration to delete
          example: "ISCC:MAIGIEGXNGY664AB"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IsccNoteDelete'
            examples:
              delete_example:
                summary: Delete declaration request
                value:
                  $schema: "http://purl.org/iscc/schema/iscc-note-delete-0.8.0.json"
                  iscc_id: "ISCC:MAIGIEGXNGY664AB"
                  timestamp: "2025-01-15T12:00:00.000Z"
                  nonce: "001d17437f53f6899ed01ecb8659118b"
                  signature:
                    version: "ISCC-SIG v1.0"
                    controller: "did:web:example.com"
                    pubkey: "z6MkmeDbeC5BecFmVnTHA5PWEBaVUrGLdB3weGE2KYnXfHso"
                    proof: "z5j9nrpPw3oYSAN4XbCvk2sUtkwrueTD6V2Y35gS1KFTode2ED2YQWokPmoXw6QBYtYEFxtAQfzBhdNyr8PMwP79G"
      responses:
        '204':
          description: No Content - Declaration successfully deleted
        '400':
          description: Bad Request - Client error (malformed request, validation errors, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid signature or not authorized to delete this declaration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found - ISCC-ID does not exist or declaration already deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity - Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    SearchLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: |
        Maximum number of matches to return (1-100). Defaults to the backend's
        own default (100) when omitted. Out-of-range values yield a 400 error.
      example: 10
  responses:
    SearchBadRequest:
      description: |
        Bad Request - malformed query, unknown top-level body key (including
        `simprints`), or out-of-range `limit` (Hub error envelope), or a
        deterministic bad-query response forwarded verbatim from the search
        backend (free-form body, not a stable contract).
      content:
        application/json:
          schema:
            anyOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                title: BackendErrorBody
                description: Backend error body forwarded verbatim (passthrough)
    SearchNotFound:
      description: |
        Not Found - similarity search is not configured on this Hub (Hub error
        envelope), or the backend reports the queried `iscc_id` reference or
        index as not found (free-form passthrough body, not a stable contract).
      content:
        application/json:
          schema:
            anyOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                title: BackendErrorBody
                description: Backend error body forwarded verbatim (passthrough)
    SearchUnprocessable:
      description: |
        Unprocessable Entity - deterministic backend validation error forwarded
        verbatim (free-form passthrough body, not a stable contract).
      content:
        application/json:
          schema:
            anyOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                title: BackendErrorBody
                description: Backend error body forwarded verbatim (passthrough)
...
  schemas:
    ErrorResponse:
      $ref: './schemas/ErrorResponse.yaml'
    DeclarationAck:
      $ref: './schemas/DeclarationAck.yaml'
    IsccDeclaration:
      $ref: './schemas/IsccDeclaration.yaml'
...
    IsccNote:
      $ref: './schemas/IsccNote.yaml'
...
    IsccQuery:
      $ref: './schemas/IsccQuery.yaml'
    IsccSearchResult:
      $ref: './schemas/IsccSearchResult.yaml'
...